Changes:    squadt:     - added diagraphica, lps2pbes and pbes2bes to list of default tools     - manual window is now centered on the main window    diagraphica:     - merged spinctrlfloat with the one in the wxe pseudo-library, the code is now       shared between ltsgraph, ltsview and diagraphica     - removed bundled tinyXML library, the one bundled with the toolset is now used     - fixed some compile warnings and a compile error (when compiling against       wxWidgets 2.6)
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@2879 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ src/libraries/wxe/spinctrlfloat.h
@@ -1,17 +1,31 @@
+// --- spinctrlfloat.h ------------------------------------------------
+// (c) 2006 -  S.W.C. Ploeger, D. Reniers  - Eindhoven Univ Technology
+// ---------------------------  *  ------------------------------------
+
+
 #ifndef SPINCTRLFLOAT_H
 #define SPINCTRLFLOAT_H
+
+#include <cmath>
 #include <string>
 #include <wx/wx.h>
 #include <wx/spinbutt.h>
+#include <wx/spinctrl.h>
 
 class wxSpinCtrlFloat : public wxPanel
 {
   public:
-    wxSpinCtrlFloat(wxWindow* parent, wxWindowID id, float p_Min, float p_Max,
-	float p_Rate, float p_Init, const wxPoint& pos = wxDefaultPosition,
-	const wxSize& size = wxDefaultSize);
+    wxSpinCtrlFloat(
+        wxWindow* parent, 
+        wxWindowID id, 
+        double p_Min, 
+        double p_Max,
+        double p_Rate, 
+        double p_Init, 
+        const wxPoint& pos = wxDefaultPosition,
+        const wxSize& size = wxDefaultSize );
 
-    float GetValue();
+    double GetValue();
     void SetValue(float v);
 protected:
     void OnEnter(wxCommandEvent & event);
@@ -19,7 +33,7 @@ protected:
     wxTextCtrl* m_TextCtrl;
     wxSpinButton* m_SpinButton;
 	
-    float m_Rate;
+    double m_Rate;
     int m_Min;
     int m_Max;
     int m_ID;
